backend/requirements.txt
py-sudoku
uvicorn
fastapi
deta